StopView split into StopView & StopForm
